+++ package.json
@@ -14,15 +14,14 @@
     "format": "prettier --write \"**/*.js\""
   },
   "dependencies": {
+    "native-base": "^2.3.2",
     "react": "16.0.0-beta.5",
     "react-native": "0.49.3"
   },
   "devDependencies": {
+    "babel-eslint": "^8.0.1",
     "babel-jest": "21.2.0",
     "babel-preset-react-native": "4.0.0",
-    "jest": "21.2.1",
-    "react-test-renderer": "16.0.0-beta.5",
-    "babel-eslint": "^8.0.1",
     "eslint": "^4.8.0",
     "eslint-config-standard": "^10.2.1",
     "eslint-config-standard-react": "^5.0.0",
@@ -32,7 +31,9 @@
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-react": "^7.4.0",
     "eslint-plugin-standard": "^3.0.1",
-    "prettier": "^1.7.4"
+    "jest": "21.2.1",
+    "prettier": "^1.7.4",
+    "react-test-renderer": "16.0.0-beta.5"
   },
   "jest": {
     "preset": "react-native"
